Feature/AX-4730 (#3634)
* add basic mac support

* Revert "add basic mac support"

This reverts commit 05a9041ef03816199727f0e7bb69676b0e437de3.

* fix qualys qid->cve, empty version exception

* yield data only if we found something

* fix qualys mapping

* handle wierd states in sa

* dont remove readme

* fix add tenable vuln

* fix version raw parsing

* impoved juniper

* fix static analysis

* Fix internal axon id

* fix shodan execution

+++ plugins/static_analysis/tests/test_static_analysis.py
@@ -29,6 +29,8 @@ def nvd_searcher():
 
     print('\nCleaning up')
     for nvd_file in glob.glob(f'{ARTIFACT_FOLDER}/*'):
+        if nvd_file == 'README.md':
+            continue
         try:
             os.remove(nvd_file)
             print(f'Removed {nvd_file}')
